WARCRAFT/requirements.txt
numpy
torch
torchvision
ray
tensorboardX
pathlib2
psutil
gym-minigrid
pillow
pandas
matplotlib
qpth
cvxpy
scipy
ax-platform
scikit-learn==0.23
